+++ megatron/core/dist_checkpointing/gpt_checkpoint_interop.py
